Reimplement QueryAsync with IAsyncEnumerable. Also cleanup up some Tasks.
Note: PagedAsync now returns IAsyncEnumerable in Items.  This means that it no longer behaves like Paged(), but then again neither does All() anymore.  Having PagedAsync return the same IEnumerable as Paged would mean that the results would have to be enumerated before they were used, resulting in a difference in behavior (Paged allows you to get the count without getting the items just by not enumerating items). This change to PagedAsync makes it a bit more difficult to deal with but I think it's the right choice.